+++ adminclient/public/index.html
@@ -43,6 +43,11 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
+    <style>
+      /* .body {
+        background-color: #aeb1a9;
+      } */
+    </style>
     <title>React App</title>
   </head>
   <body>
