+++ adminclient/src/Components/Student Courses/StudentCourses.jsx
@@ -1,12 +1,12 @@
 import React from "react";
 import MaterialTable from "material-table";
 import { Button, Container } from "@material-ui/core";
-import CreateIcon from '@material-ui/icons/Create';
+import CreateIcon from "@material-ui/icons/Create";
 // import ClassroomActions from "./ClassroomActions";
 import { useHistory } from "react-router";
 
 const StudentCourses = () => {
-  const history = useHistory()
+  const history = useHistory();
   const data = [
     {
       DclassName: "phy",
@@ -69,23 +69,43 @@ const StudentCourses = () => {
       LecturesComplete: "2",
     },
   ];
-  
-  const handleManageCourse = () =>{
-    history.push("/student-courses/manage")
-  }
+
+  const handleManageCourse = () => {
+    history.push("/student-courses/manage");
+  };
   const Columns = [
     { title: "Name", field: "DclassName" },
     { title: "Teacher", field: "teacher" },
-    { title: "Student", field: "student",Style: {border: "1px solid black"} },
-    { title: "Lectures Completed", field: "LecturesComplete",render: (row) => <div>Hello</div> },
-    {title: "Actions",field: "name",render: (row) => 
-    <div>
-        <Button variant="contained" style={{backgroundColor:"#17a2b8",padding:"2%",color:"white",fontSize:"small"}} 
-        onClick={handleManageCourse}
-        >
-            <CreateIcon/> Manage Student Course
-        </Button>
-    </div>}
+    {
+      title: "Student",
+      field: "student",
+      Style: { border: "0.5px solid #ccc" },
+    },
+    {
+      title: "Lectures Completed",
+      field: "LecturesComplete",
+      render: (row) => <div>Hello</div>,
+    },
+    {
+      title: "Actions",
+      field: "name",
+      render: (row) => (
+        <div>
+          <Button
+            variant="contained"
+            style={{
+              backgroundColor: "#17a2b8",
+              padding: "2%",
+              color: "white",
+              fontSize: "small",
+            }}
+            onClick={handleManageCourse}
+          >
+            <CreateIcon /> Manage Student Course
+          </Button>
+        </div>
+      ),
+    },
   ];
   return (
     <div>
@@ -95,23 +115,23 @@ const StudentCourses = () => {
           title="Students"
           data={data}
           columns={Columns}
-          options={{ exportButton: true,
+          options={{
+            exportButton: true,
             border: true,
             headerStyle: {
-                // backgroundColor: "pink"
-                border: "1px solid black"
-
+              // backgroundColor: "pink"
+              border: "0.5px solid #ccc",
             },
             cellStyle: {
-                border: "1px solid black"
-            }
+              border: "0.5px solid #ccc",
+            },
           }}
-        //   actions ={[
-        //       {
-                  
-        //       }
-        //   ]}
-          style={{border: "1px solid black"}}
+          //   actions ={[
+          //       {
+
+          //       }
+          //   ]}
+          style={{ border: "0.5px solid #ccc" }}
         />
       </Container>
     </div>
