+++ adminclient/src/Helpers/attendanceData.js
@@ -1,0 +1,64 @@
+export const attendanceData = () => {
+  return [
+    {
+      DclassName: "phy",
+      teacher: "manisha",
+      student: "0",
+      LecturesComplete: "1",
+    },
+    {
+      DclassName: "phy",
+      teacher: "manisha",
+      student: "0",
+      LecturesComplete: "1",
+    },
+    {
+      DclassName: "phy",
+      teacher: "manisha",
+      student: "0",
+      LecturesComplete: "1",
+    },
+    {
+      DclassName: "phy",
+      teacher: "manisha",
+      student: "0",
+      LecturesComplete: "1",
+    },
+    {
+      DclassName: "phy",
+      teacher: "manisha",
+      student: "5",
+      LecturesComplete: "1",
+    },
+    {
+      DclassName: "phy",
+      teacher: "manisha",
+      student: "0",
+      LecturesComplete: "1",
+    },
+    {
+      DclassName: "phy",
+      teacher: "manisha",
+      student: "0",
+      LecturesComplete: "1",
+    },
+    {
+      DclassName: "phy",
+      teacher: "priya",
+      student: "0",
+      LecturesComplete: "2",
+    },
+    {
+      DclassName: "phy",
+      teacher: "isha",
+      student: "0",
+      LecturesComplete: "1",
+    },
+    {
+      DclassName: "sci",
+      teacher: "manisha",
+      student: "0",
+      LecturesComplete: "2",
+    },
+  ];
+};
