+++ adminclient/src/Pages/OneToOne.jsx
@@ -1,12 +1,12 @@
 import React from "react";
 import MaterialTable from "material-table";
 import { Button, Container } from "@material-ui/core";
-import CreateIcon from '@material-ui/icons/Create';
+import CreateIcon from "@material-ui/icons/Create";
 import ClassroomActions from "./ClassroomActions";
 import { useHistory } from "react-router";
 
 const OneToOne = () => {
-  const history = useHistory()
+  const history = useHistory();
   const data = [
     {
       DclassName: "phy",
@@ -69,28 +69,45 @@ const OneToOne = () => {
       LecturesComplete: "2",
     },
   ];
-  
-  const handleAddNewClassroom = () =>{
-    history.push("/classroom/create")
-  }
+
+  const handleAddNewClassroom = () => {
+    history.push("/classroom/create");
+  };
   const Columns = [
     { title: "Name", field: "DclassName" },
     { title: "Teacher", field: "teacher" },
     { title: "Student", field: "student" },
-    { title: "Lectures Completed", field: "LecturesComplete",render: (row) => <div>Hello</div> },
-    {title: "Actions",field: "name",render: (row) => <div><ClassroomActions/></div>}
+    {
+      title: "Lectures Completed",
+      field: "LecturesComplete",
+      render: (row) => <div>Hello</div>,
+    },
+    {
+      title: "Actions",
+      field: "name",
+      render: (row) => (
+        <div>
+          <ClassroomActions />
+        </div>
+      ),
+    },
   ];
   return (
     <div>
       <Container>
         {" "}
-        <Button variant="contained" color="secondary" onClick={handleAddNewClassroom}>Add new Classroom</Button>
+        <Button
+          variant="contained"
+          color="secondary"
+          onClick={handleAddNewClassroom}
+        >
+          Add new Classroom
+        </Button>
         <MaterialTable
           title=""
           data={data}
           columns={Columns}
-          options={{ exportButton: true,
-          }}
+          options={{ exportButton: true }}
           // actions ={
           //   <Button>Hi</Button>
           // }
